Solo enteros

+++ app/Models/Feed.php
@@ -113,7 +113,7 @@ class Feed extends Model
                 'description' => strip_tags(isset($item['description']) ? (string)$item['description'] : ''),
                 'publication_date' => isset($item['pubDate']) ? (new Carbon((string)$item['pubDate']))->toDateTimeString() : null,
                 'media_url' => isset($item['enclosure']['@attributes']['url']) ? (string)$item['enclosure']['@attributes']['url'] : '',
-                'duration' => isset($item['enclosure']['@attributes']['length']) ? (string)$item['enclosure']['@attributes']['length'] : 0,
+                'duration' => (int)isset($item['enclosure']['@attributes']['length']) ? (string)$item['enclosure']['@attributes']['length'] : 0,
             ];
 
             if (!$data['media_url']) {
